requirements.txt
fastapi
uvicorn
databases
asyncpg
python-dotenv
sqlalchemy
psycopg2-binary